[wnsxor1993#86] refactor: IssueDetailResponse의 내부 클래스를 분리
- `IssueDetailResponse`의 내부 클래스인 `IssueAssigneeDto`를 따로 분리
- `IssueDetailResponse`의 내부 클래스인 `CommentListElement`를 따로 분리
